+++ .vscode/settings.json
@@ -7,6 +7,7 @@
         "editor.defaultFormatter": "ms-python.black-formatter"
     },
     "python.analysis.extraPaths": [
-        "./packages/processor"
+        "./packages/processor",
+        "./packages/parse"
     ]
 }
