+++ packages/indexer/docker-compose.tests.yaml
@@ -12,7 +12,10 @@ x-common-variables:
   - GOOGLE_CLOUD_PROJECT=${GCP_PROJECT_ID}
 services:
   indexer-test:
-    image: indexer-dev
+    image: indexer
+    build:
+      context: .
+      dockerfile: Dockerfile
     command: pytest --junitxml=pytest.xml --cov="indexer" --cov-report=xml --cov-report term-missing --cov-fail-under=90 -vvv
     depends_on:
       - test_postgres
@@ -20,9 +23,6 @@ services:
     volumes:
       - .:/indexer
       - $GOOGLE_APPLICATION_CREDENTIALS:/indexer/creds.json
-    build:
-      context: .
-      dockerfile: Dockerfile
     container_name: indexer-test
   test_postgres:
     image: postgres:14
