+++ packages/indexer/pyproject.toml
@@ -15,7 +15,7 @@ pandas = "^2.0"
 deepdiff = "^6.3.0"
 google-cloud-storage = "^2.8.0"
 aiofiles = "^23.1.0"
-parse = { path = "./../parse", develop = true }
+parse = "^0.1.0"
 pytest = "^7.2.2"
 pytest-asyncio = "^0.20.3"
 pytest-cov = "^4.0.0"
