+++ packages/parse/README.md
@@ -1,0 +1,5 @@
+# Parse
+
+This package contains all of the parsing logic for the raw blockchain data.
+
+This is separate from our indexer so other packages can use the logic
