+++ packages/parse/docker-compose.tests.yaml
@@ -1,0 +1,13 @@
+version: '3'
+services:
+  parse-test:
+    container_name: parse
+    build:
+      dockerfile: Dockerfile
+    image: parser
+    command: pytest --junitxml=pytest.xml --cov="parse" --cov-report=xml --cov-report term-missing --cov-fail-under=90 -vvv
+networks:
+  default:
+    name: sn-mono-network
+    driver: bridge
+    external: true
