+++ packages/parse/parse/exceptions.py
@@ -1,0 +1,6 @@
+class ParseError(BaseException):
+    pass
+
+
+class BlockPrimaryKeyNotDefinedError(ParseError):
+    pass
