+++ packages/parse/pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "parse"
+version = "0.1.0"
+description = ""
+authors = ["Ethan Leifer <ethanleifer@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "3.10.*"
+pytest = "^7.3.1"
+pytest-cov = "^4.0.0"
+pytest-mock = "^3.10.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+pythonpath="."
