+++ packages/parse/pytest.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="5" time="0.150" timestamp="2023-05-20T12:55:45.799940" hostname="Ethans-MacBook-Pro-1183.local"><testcase classname="tests.processor_test" name="test_parse_individual_log_event" time="0.001" /><testcase classname="tests.processor_test" name="test_parse_tx_logs" time="0.000" /><testcase classname="tests.processor_test" name="test_parse_tx_log_error" time="0.000" /><testcase classname="tests.processor_test" name="test_parsing" time="0.033" /><testcase classname="tests.processor_test" name="test_parse_tx_before_block_error" time="0.021" /></testsuite></testsuites>
