+++ packages/parse/tests/test_data/log_data.json
@@ -1,0 +1,48 @@
+[
+    {
+        "events": [
+            {
+                "type": "message",
+                "attributes": [
+                    {
+                        "key": "action",
+                        "value": "/secret.compute.v1beta1.MsgExecuteContract"
+                    },
+                    {
+                        "key": "module",
+                        "value": "compute"
+                    },
+                    {
+                        "key": "sender",
+                        "value": "secret1rjml9f6ma7rwgqq9ud8e0aw2arhhj4v4hhzc4k"
+                    },
+                    {
+                        "key": "contract_address",
+                        "value": "secret1266jqzsyw98g3v8cz5cyhw2s9kwhtmtdnr0898"
+                    }
+                ]
+            },
+            {
+                "type": "wasm",
+                "attributes": [
+                    {
+                        "key": "contract_address",
+                        "value": "secret1266jqzsyw98g3v8cz5cyhw2s9kwhtmtdnr0898"
+                    },
+                    {
+                        "key": "CjABfNUNmW/ZmkxacGL/yqDywwk+ew==",
+                        "value": "Odng8iw3EBn0t4v+EwnnQFcBICMlbaebFvusZxxENqz6PUY5"
+                    },
+                    {
+                        "key": "Yxlzf/42hUhOklVo87eSe1O5FyaVnXi+Hw==",
+                        "value": "/X5CRFEbnjNX1RMf69r19XiYryGJLh/cV4swLLorrKJIJnyB4qrkd5aPrlAJEHzpSK1bjpgdKBKAtG68Ag=="
+                    },
+                    {
+                        "key": "contract_address",
+                        "value": "secret1rgm2m5t530tdzyd99775n6vzumxa5luxcllml4"
+                    }
+                ]
+            }
+        ]
+    }
+]
