--- packages/processor/pyproject.toml
@@ -1,14 +1,0 @@
-[tool.poetry]
-name = "processor"
-version = "0.1.0"
-description = ""
-authors = ["Ethan Leifer <ethanleifer@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "3.10.*"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
