+++ poetry/projects.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# all python packages, in topological order
+PROJECTS='api cli dags dashboard indexer parse'
